WriteBuffer accepts an io.Writer rather than a *os.File
This allows a caller to build an iso image in memory.

Since io.Writer only includes a Write() method avoid the use of Truncate() and
Seek() by simply writing a zeroed byte slice of the correct length instead of
writing a random poem of arbitrary length, truncating (which may actually be
zero extending the file if the desired number of bytes happens to be fewer than
are contained in The Raven) and seeking.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>